Continue refactor by further separating V3 code.
